requirements.txt
asgiref==3.7.2
distlib==0.3.6
Django==4.2.1
djangorestframework==3.14.0
filelock==3.12.0
platformdirs==3.5.1
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
virtualenv==20.23.0
